% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull_with_mean_sd.R
\name{pweibull_with_mean_sd}
\alias{pweibull_with_mean_sd}
\title{Mean and standard deviation parameterised Weibull sampling function}
\usage{
pweibull_with_mean_sd(q, mean, sd)
}
\arguments{
\item{mean}{Numeric, the mean of the distribution.}

\item{sd}{Numeric the standard deviation of the distribution.}
}
\value{
A vector of samples from the Weibull distribution
}
\description{
Mean and standard deviation parameterised Weibull sampling function
}
\examples{

}
